pyproject.toml
[tool.poetry]
name = "financial-news-sentiment"
version = "0.1.0"
description = "A production-ready financial news sentiment analysis system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "financial_news_sentiment", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.5.2"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
pyyaml = "^6.0.1"
httpx = ">=0.27"
numpy = ">=1.26.0,<2.0.0"
torch = "^2.0.0"
transformers = "^4.31.0"
langchain = "^0.1.12"
langchain-core = "^0.1.30"
langchain-community = "^0.0.28"
lxml = {extras = ["html", "html_clean"], version = "^6.0.0"}
newspaper3k = "^0.2.8"
yfinance = "^0.2.28"
scikit-learn = "^1.4.0"
ollama = "^0.5.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"
flake8 = "^6.1.0"
bandit = "^1.7.5"
pre-commit = "^3.3.3"
pdoc = "^14.0.0"
interrogate = "^1.5.0"
pdoc3 = "^0.11.6"
types-requests = "^2.32.4.20250809"
types-pyyaml = "^6.0.12.20250822"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fins = "financial_news_sentiment.cli:main"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-magic = true
ignore-semiprivate = true
ignore-private = true
ignore-property-decorators = true
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = true
fail-under = 95
verbose = 2
